parse watchdogsec from systemd file
