+++ configuration/init/systemd/tedge-mapper-az.service
@@ -7,7 +7,6 @@ User=tedge-mapper
 ExecStart=/usr/bin/tedge_mapper az
 Restart=on-failure
 RestartPreventExitStatus=255
-WatchdogSec=15
 
 [Install]
 WantedBy=multi-user.target
