+++ configuration/init/systemd/tedge-mapper-c8y.service
@@ -7,7 +7,6 @@ User=tedge-mapper
 ExecStart=/usr/bin/tedge_mapper c8y
 Restart=on-failure
 RestartPreventExitStatus=255
-WatchdogSec=15
 
 [Install]
 WantedBy=multi-user.target
