+++ configuration/init/systemd/tedge-watchdog.service
@@ -1,11 +1,11 @@
 [Unit]
-Description=tedge-monitor: Checks the health of all thin-edge.io services
+Description=tedge-watchdog: Checks the health of all thin-edge.io services
 After=syslog.target network.target mosquitto.service
 StartLimitIntervalSec=0
 
 [Service]
 Type=notify
-ExecStart=/usr/bin/tedge_monitor
+ExecStart=/usr/bin/tedge_watchdog
 Restart=on-failure
 
 [Install]
