+++ crates/core/tedge_watchdog/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "tedge_monitor"
+name = "tedge_watchdog"
 version = "0.6.0"
 authors = ["thin-edge.io team <info@thin-edge.io>"]
 edition = "2021"
 rust-version = "1.58.1"
 license = "Apache-2.0"
-description = "tedge_monitor checks the health of all the thin-edge.io components/services."
+description = "tedge_watchdog checks the health of all the thin-edge.io components/services."
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -15,6 +15,7 @@ mqtt_channel = { path = "../../common/mqtt_channel" }
 nanoid = "0.4.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+freedesktop_entry_parser = "1.3.0"
 tedge_config = { path = "../../common/tedge_config" }
 thiserror ="1.0.30"
 tokio = { version = "1.12", features = ["sync",  "time"] }
